[ILUVATAR_GPU] Support for iluvatar_gpu

+++ ppocr/data/imaug/operators.py
@@ -230,7 +230,7 @@ class DetResizeForTest(object):
             # img, shape = self.resize_image_type1(img)
             img, [ratio_h, ratio_w] = self.resize_image_type1(img)
         data["image"] = img
-        data["shape"] = np.array([src_h, src_w, ratio_h, ratio_w])
+        data["shape"] = np.array([src_h, src_w, ratio_h, ratio_w]).astype(np.float32)
         return data
 
     def image_padding(self, im, value=0):
